Chat_with_CSV/requirements.txt
langchain_experimental
pandas==2.2.2
langchain==0.2.5
pypdf2==3.0.1
python-dotenv==1.0.1
streamlit==1.35.0
langchain_community==0.2.5
faiss-cpu==1.8.0 
openai==1.34.0
tiktoken==0.7.0
langchain_openai==0.1.8